Refactored some functions
