+++ functions/package.json
@@ -34,7 +34,7 @@
     "eslint-config-google": "^0.14.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-import": "^2.25.4",
-    "firebase-functions-test": "^3.1.0",
+    "firebase-functions-test": "^3.4.0",
     "prettier": "^3.4.2",
     "prettier-eslint": "^16.3.0",
     "typescript": "^4.9.0"
