+++ functions/src/@types/firebase-data.ts
@@ -1,0 +1,1 @@
+export {};
