+++ functions/src/commands/register.ts
@@ -1,8 +1,10 @@
-import type { AppCommand } from "../utils/types";
+import type { AppCommand } from "../@types/discord-custom";
 import * as aoe4world from "../utils/aoe4world";
 import { REGISTER } from "../constants/command-list";
-import { doesGuildMatch } from "../utils/request-processing";
 import { GAMES } from "../constants/games";
+import { validateCredentials, validateGame } from "../lib/helpers/validator";
+import { getDiscordUserId } from "../lib/helpers/getters";
+import { getFirestore } from "firebase-admin/firestore";
 // TODO: @ path resolution is not working when build completes
 
 // https://interactions-zhgogk6p4a-uc.a.run.app
@@ -11,53 +13,28 @@ const MAX_IDS = 5;
 
 export default {
   ...REGISTER,
-  callback: async ({ interaction, db }) => {
-    if (
-      process.env.NODE_ENV === "production" &&
-      interaction.channel.id !== process.env.LEADERBOARD_CHANNEL_ID
-    ) {
-      return {
-        content: "This command is only available in the leaderboard channel"
-      };
-    }
-
-    if (!interaction.guild_id || !doesGuildMatch(interaction.guild_id)) {
-      return {
-        content: "This command is only available in the ICRS Discord server"
-      };
-    }
-
-    const discordUserId = interaction?.member?.user?.id;
-    if (!discordUserId) {
-      return {
-        content: "Could not find your user ID"
-      };
-    }
-
-    console.log(interaction);
+  callback: async ({ interaction }) => {
+    validateCredentials(interaction);
+    const discordUserId = getDiscordUserId(interaction);
 
     const game = interaction.data?.options?.[0].value as string;
-    if (!game || !Object.values(GAMES).includes(game as GAMES)) {
-      return {
-        content: "Please select a game we support"
-      };
-    }
+    validateGame(game);
 
     const id = interaction.data?.options?.[1].value as string;
     if (!id) {
-      return {
-        content: "Please provide a valid ID"
-      };
+      throw new Error("Please provide a valid ID");
     }
+
+    const db = getFirestore();
 
     let existingIds = {};
     const userCollection = db.collection("users");
     const userDoc = await userCollection.doc(`${discordUserId}`).get();
     existingIds = userDoc.exists ? userDoc.data()?.[game] : {};
     if (Object.keys(existingIds).length >= MAX_IDS) {
-      return {
-        content: "You have reached the maximum number of IDs for this game"
-      };
+      throw new Error(
+        `You have already registered ${MAX_IDS} profiles. Please deregister one before registering a new one.`
+      );
     }
 
     let player = null;
@@ -66,17 +43,15 @@ export default {
     }
 
     if (!player) {
-      return {
-        content: "Could not find player with the provided ID"
-      };
+      throw new Error(`Could not find a player with the given ID : ${id}`);
     }
     const playerRating = (
       await aoe4world.getSoloLeaderboard([player.profile_id])
     )[player.profile_id];
     if (!playerRating) {
-      return {
-        content: "Could not find rating for the player"
-      };
+      throw new Error(
+        `Could not find the rating for the player with the given ID : ${id}`
+      );
     }
 
     await db
